Estimate Escapement - LGD

Combines estimates of total escapement from STADEM with transition probabilities from DABOM to generate estimates of escapement above various detection sites. Currently only works for wild fish.

calcBranchEscape_GRA(
  dabom_mod = NULL,
  stadem_mod = NULL,
  stadem_param_nm = "X.new.wild",
  bootstrap_samp = 2000
)

Arguments

dabom_mod

The result of DABOM. An MCMC.list, where the detection parameter names all end with "_p".

stadem_mod

The result of STADEM. An MCMC.list, or jagsUI object that can be coerced to an MCMC.list.

stadem_param_nm

Character vector of length one, with the name of the parameter corresponding to total escapement past Lower Granite Dam.

bootstrap_samp

The number of samples to be drawn from the posteriors of the STADEM model and the DABOM model.

Author

Kevin See

Examples

#calcBranchEscape_GRA()