Estimate Escapement - LGD

Combines estimates of total escapement from STADEM with transition probabilities from DABOM to generate estimates of escapement above various detection sites. Currently only works for wild fish.

calcTribEscape_GRA(
  dabom_mod = NULL,
  stadem_mod = NULL,
  stadem_param_nm = "X.new.wild",
  bootstrap_samp = 2000,
  parent_child = NULL,
  summ_results = T,
  cred_int_prob = 0.95
)

Arguments

dabom_mod
The result of DABOM. An MCMC.list, where the detection parameter names all end with "_p".

stadem_mod
The result of STADEM. An MCMC.list, or jagsUI object that can be coerced to an MCMC.list.

stadem_param_nm
Character vector of length one, with the name of the parameter corresponding to total escapement past Lower Granite Dam.

bootstrap_samp
The number of samples to be drawn from the posteriors of the STADEM model and the DABOM model.

parent_child
data frame with at least `parent` and `child` columns. Can be created with `buildParentChild()` function in the `PITcleanr` package.

summ_results
Should the resulting posteriors be summarized or returned as posterior samples? Default value is TRUE.

cred_int_prob
A numeric scalar in the interval (0,1) giving what highest posterior density portion of the posterior the credible interval should cover. The default value is 95%.

Author

Kevin See

Examples

#calcTribEscape_GRA()